+++ packages/main/src/enums/MediaGalleryLayout.ts
@@ -1,0 +1,21 @@
+// Generated file - do not change manually!
+
+/**
+ * Defines the layout type of the thumbnails list of the `MediaGallery` component.
+ */
+export enum MediaGalleryLayout {
+  /**
+   * The layout is determined automatically.
+   */
+  Auto = 'Auto',
+
+  /**
+   * Displays the layout as a horizontal split between the thumbnails list and the selected image.
+   */
+  Horizontal = 'Horizontal',
+
+  /**
+   * Displays the layout as a vertical split between the thumbnails list and the selected image.
+   */
+  Vertical = 'Vertical'
+}