+++ packages/main/src/enums/ZIndex.ts
@@ -1,7 +1,0 @@
-export enum ZIndex {
-  DialogOverlay = 950,
-  Dialog = 1000,
-  MessageBox = 1100,
-  ResponsivePopover = 1200,
-  InputModal = 1300
-}