+++ packages/main/src/webComponents/Link/index.tsx
@@ -9,6 +9,21 @@ import '@ui5/webcomponents/dist/Link.js';
 
 interface LinkAttributes {
   /**
+   * An object of strings that defines several additional accessibility attribute values for customization depending on the use case. It supports the following fields:
+   *
+   * *   `expanded`: Indicates whether the anchor element, or another grouping element it controls, is currently expanded or collapsed. Accepts the following string values:
+   *     *   `true`
+   *     *   `false`
+   *
+   * *   `hasPopup`: Indicates the availability and type of interactive popup element, such as menu or dialog, that can be triggered by the anchor element. Accepts the following string values:
+   *     *   `Dialog`
+   *     *   `Grid`
+   *     *   `ListBox`
+   *     *   `Menu`
+   *     *   `Tree`
+   */
+  accessibilityAttributes?: Record<string, unknown>;
+  /**
    * Receives id(or many ids) of the elements that label the input
    */
   accessibleNameRef?: string;
@@ -74,7 +89,7 @@ export interface LinkPropTypes extends LinkAttributes, Omit<CommonProps, 'onClic
  */
 const Link = withWebComponent<LinkPropTypes, LinkDomRef>(
   'ui5-link',
-  ['accessibleNameRef', 'design', 'href', 'target', 'wrappingType'],
+  ['accessibilityAttributes', 'accessibleNameRef', 'design', 'href', 'target', 'wrappingType'],
   ['disabled'],
   [],
   ['click']